API/requirements.txt
asyncpg==0.27.0
fastapi==0.88.0
pydantic==1.10.2
pytest==7.2.0
PyYAML==6.0
uvicorn==0.20.0
